fix: floats and allow complete expenses request arguments

+++ packages/sdk/src/helpers/xml.ts
@@ -73,7 +73,7 @@ export function parseXML(xml: string): unknown {
             }
 
             if (Number.isNaN(+tagValue) === false) {
-                return parseInt(tagValue)
+                return parseFloat(tagValue)
             }
 
             return tagValue
